fix(ffe-message-box-react): Update docs
Fixes an error in the documentation where the code-snippet on
the designguide was pointing to the wrong export names.

Also adds a section for props & methods for each box and removes
documentation for the base message class, which was never exported

Fixes #252
