+++ packages/ffe-message-box-react/src/ErrorMessage.md
