+++ packages/ffe-message-box-react/src/InfoMessage.md
