+++ packages/ffe-message-box-react/src/TipsMessage.md
