+++ packages/ffe-message-box-react/src/index.js
@@ -1,7 +1,7 @@
-import SuccessMessage from './SuccessMessageBox';
-import ErrorMessage from './ErrorMessageBox';
-import InfoMessage from './InfoMessageBox';
-import TipsMessage from './TipsMessageBox';
+import SuccessMessage from './SuccessMessage';
+import ErrorMessage from './ErrorMessage';
+import InfoMessage from './InfoMessage';
+import TipsMessage from './TipsMessage';
 import { InfoMessageList, InfoMessageListItem } from './InfoMessageList';
 
 export {
